requirements.txt
torch
numpy
onnx
onnxruntime
onnxscript
pytest
flake8

hydra-core
omegaconf
mlflow
